requirements.txt
Django == 4.2.3
psycopg2-binary
graphene-django
strawberry-graphql == 0.198.0
uvicorn
strawberry-django-auth
pytest-django
mixer==7.2.1
